Add and update files via upload

+++ blog/blog.py
@@ -1,3 +1,7 @@
+#!/usr/bin/env python
+
+"""blog.py: Controls multi-user blog functionality using Google App Engine"""
+
 import os
 import re
 import jinja2
@@ -6,19 +10,20 @@ import hashlib
 import hmac
 import string
 import random
+import time
 
 from google.appengine.ext import db
 
 template_dir = os.path.join(os.path.dirname(__file__), 'templates')
 jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(template_dir),
                                 autoescape = True)
-secret = 'asdlfaje09349q8[3fe38y33ru37y7;fu'
+secret = "asdlfaje09349q8[3fe38y33ru37y7;fu"
 
-USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
+USER_RE = re.compile(r'^[a-zA-Z0-9_-]{3,20}$')
 def valid_username(username):
     return username and USER_RE.match(username)
 
-PASS_RE = re.compile(r"^.{3,20}$")
+PASS_RE = re.compile(r'^.{3,20}$')
 def valid_password(password):
     return password and PASS_RE.match(password)
 
@@ -48,6 +53,7 @@ def valid_pw(name, pw, h):
     return h == make_pw_hash(name, pw, salt)
 
 class Handler(webapp2.RequestHandler):
+    """Base handler that defines rendering and cookie functions for child handlers."""
     def write(self, *a, **kw):
         self.response.out.write(*a, **kw)
 
@@ -78,37 +84,43 @@ class Handler(webapp2.RequestHandler):
         uid = self.read_secure_cookie('user_id')
         self.user = uid and User.by_id(int(uid))
 
+
 class Blog(db.Model):
+    """Handles creation of blog entities."""
     subject = db.StringProperty(required = True)
     content = db.TextProperty(required = True)
     created = db.DateTimeProperty(auto_now_add = True)
     #New columns. Changes required.
-    user_id = db.StringProperty(required = True)
+    user_id = db.StringProperty()
     likes = db.IntegerProperty(default=0)
     likers = db.StringListProperty()
     dislikes = db.IntegerProperty(default = 0)
 
+
 class Comment(db.Model):
+    """Handles creation of comment entities."""
     content = db.TextProperty(required = True)
     created = db.DateTimeProperty(auto_now_add = True)
     user_id = db.IntegerProperty(required = True)
+    blog_id = db.IntegerProperty()
     blog = db.ReferenceProperty(Blog, collection_name='blog_comments')
 
+
 class User(db.Model):
+    """Handles creation of comment entities."""
     name = db.StringProperty(required = True)
     pw_hash = db.StringProperty(required = True)
     email = db.StringProperty()
 
     #classmethod means it acts on the class itself, not an instance
+    #this code is from the homework solutions from the lesson
     @classmethod
     def by_id(cls, uid):
         return cls.get_by_id(uid)
-        #was User.get_by_id(uid)
 
     @classmethod
     def by_name(cls, name):
         u = cls.all().filter('name =', name).get()
-        #was User.all()....
         return u
 
     @classmethod
@@ -122,14 +134,15 @@ class User(db.Model):
         if u and valid_pw(name, pw, u.pw_hash):
             return u
 
-class Signup(Handler):
 
+class Signup(Handler):
+    """Defines the form validation for the register handler."""
     def get(self):
         author = self.read_secure_cookie('user_id')
         if author:
-            self.render("signup.html", author=author)
+            self.render('signup.html', author=author)
         else:
-            self.render("signup.html")
+            self.render('signup.html')
 
     def post(self):
         have_error = False
@@ -164,7 +177,10 @@ class Signup(Handler):
     def done(self, *a, **kw):
         raise NotImplementedError
 
+
 class Register(Signup):
+    """Inherits from signup and adds new users
+     to the database while also assigning cookies."""
     def done(self):
         u = User.by_name(self.username)
         if u:
@@ -174,11 +190,12 @@ class Register(Signup):
             u = User.register(self.username, self.password, self.email)
             u.put()
 
-            #set cookie
             self.login(u)
             self.redirect('/welcome')
 
+
 class Login(Handler):
+    """Handles user login and setting cookies."""
     def get(self):
         author = self.read_secure_cookie('user_id')
         if author:
@@ -197,21 +214,27 @@ class Login(Handler):
             msg = 'Invalid Login'
             self.render('login.html', error = msg)
 
+
 class Logout(Handler):
+    """Handles user logout by deleting cookies."""
     def get(self):
         self.logout()
         self.redirect('/signup')
 
+
 class HomePage(Handler):
+    """Handles the base template for the blog app."""
     def get(self):
         author = self.read_secure_cookie('user_id')
         msg = self.request.get('msg')
         if author:
-            self.render("base.html", author = author, msg = msg)
+            self.render('base.html', author = author, msg = msg)
         else:
-            self.render("base.html")
+            self.render('base.html')
+
 
 class Welcome(Handler):
+    """Renders welcome page after signup or login."""
     def get(self):
         if self.user:
             user_id = self.read_secure_cookie('user_id')
@@ -219,39 +242,39 @@ class Welcome(Handler):
         else:
             self.redirect('/signup')
 
-#Shows ten newest blogs
+
 class MainBlogPage(Handler):
+    """Renders ten most recently created blog posts."""
     def get(self):
         blogs = db.GqlQuery("SELECT * FROM Blog ORDER BY created DESC LIMIT 10")
         user_id = self.read_secure_cookie('user_id')
-        self.render("blog.html", blogs = blogs, user_id = user_id)
+        self.render('blog.html', blogs = blogs, user_id = user_id)
 
-#Goes to a page with the new blog displayed
+
 class BlogPage(Handler):
+    """Renders page for a newly created blog post.
+    Also handles the form submissions for blog posts."""
     def get(self, blog_id):
-        #key = db.Key.from_path('Blog', int(blog_id))
-        #blog = db.get(key)
         blog = Blog.get_by_id(int(blog_id))
-
         if not blog:
             self.error(404)
             return
-        #comments = db.GqlQuery("SELECT * FROM Comment ORDER BY created DESC ")
-        #comments = []
-        #for comment in blog.blog_comments:
-            #comments.append(comment.content)
-
 
         author = self.read_secure_cookie('user_id')
         if author:
-            self.render("permalink.html", blog=blog, author = author)
+            self.render('permalink.html', blog=blog, author = int(author))
         else:
-            self.render("permalink.html", blog=blog)
+            self.render('permalink.html', blog=blog)
 
+    #Checks for type of submission and follows up with the appropriate response.
     def post(self, blog_id):
         user_id = self.read_secure_cookie('user_id')
+        blog = Blog.get_by_id(int(blog_id))
+        if not blog:
+            self.error(404)
+            return
+
         if self.request.get('delete'):
-            blog = Blog.get_by_id(int(blog_id))
             if blog.user_id == user_id:
                 blog.delete()
                 msg = "Blog Deleted"
@@ -259,72 +282,67 @@ class BlogPage(Handler):
             else:
                 self.redirect('/login')
         elif self.request.get('edit'):
-            blog = Blog.get_by_id(int(blog_id))
-            user_id = self.read_secure_cookie('user_id')
-            self.redirect('/blog/edit/%s' % blog_id)
+            if blog.user_id == user_id:
+                self.redirect('/blog/edit/%s' % blog_id)
+            else:
+                self.redirect('/login')
         elif self.request.get('like'):
-            blog = Blog.get_by_id(int(blog_id))
-            if not user_id in blog.likers:
-                blog.likes += 1
-                blog.likers.append(user_id)
-                blog.put()
-                self.redirect('/blog')
+            if user_id:
+                if not user_id in blog.likers:
+                    blog.likes += 1
+                    blog.likers.append(user_id)
+                    blog.put()
+                    time.sleep(0.1)
+                    self.redirect('/blog')
+                else:
+                    self.redirect('/blog')
             else:
-                self.redirect('/blog')
+                self.redirect('/login')
         elif self.request.get('dislike'):
-            blog = Blog.get_by_id(int(blog_id))
-            if not user_id in blog.likers:
-                blog.dislikes += 1
-                blog.likers.append(user_id)
-                blog.put()
-                self.redirect('/blog')
+            if user_id:
+                if not user_id in blog.likers:
+                    blog.dislikes += 1
+                    blog.likers.append(user_id)
+                    blog.put()
+                    time.sleep(0.1)
+                    self.redirect('/blog')
+                else:
+                    self.redirect('/blog')
             else:
-                self.redirect('/blog')
+                self.redirect('/login')
         elif self.request.get('view'):
             self.redirect('/blog/%s'%blog_id)
-        # elif self.request.get('post-comment'):
-        #     content=""
-        #     if self.request.get('comment_text'):
-        #         content = self.request.get('comment_text')
-        #     c = Comment(content=content, user_id=int(user_id), blog_id=int(blog_id))
-        #     c.put()
-        #     self.redirect('/blog/%s'%blog_id)
         elif self.request.get('comment'):
-             #blog = Blog.get_by_id(int(blog_id))
-             #user_id = self.read_secure_cookie('user_id')
-             self.redirect('/comment/%s'%blog_id)
-        elif self.request.get('delete-coment'):
-            comment = self.request.get('c')
-            if c.user_id == user_id:
-                comment.delete()
-                self.redirect('/blog/%s'%blog_id)
+            if user_id:
+                self.redirect('/comment/%s'%blog_id)
             else:
-                self.redirect('/blog/%s'%blog_id)
-        elif self.request.get('edit-comment'):
-            content = self.request.get("content")
-            self.render("comment.html", content = content)
-
-
+                self.redirect('/login')
+                
 
 class CommentHandler(Handler):
+    """Handles creation of comment entities and form validation."""
     def render_newcomment(self, content="", error=""):
         author = self.read_secure_cookie('user_id')
         if author:
-            self.render("comment.html", content=content, error=error, author=author)
+            self.render('comment.html', content=content, error=error, author=author)
         else:
-            self.render("comment.html", content=content, error=error)
+            self.render('comment.html', content=content, error=error)
 
     def get(self, blog_id):
         self.render_newcomment()
 
     def post(self, blog_id):
-        content = self.request.get("content")
+        content = self.request.get('content')
         user_id = self.read_secure_cookie('user_id')
         blog = Blog.get_by_id(int(blog_id))
+        if not blog:
+            self.error(404)
+            return
 
         if content:
-            c = Comment(content = content, user_id = int(user_id), blog=blog)
+            c = Comment(content = content, user_id = int(user_id), blog_id=int(blog_id), blog=blog)
             c.put()
+            time.sleep(0.1)
             link_id = blog.key().id()
             self.redirect('/blog/%s' % link_id)
         else:
@@ -332,21 +350,81 @@ class CommentHandler(Handler):
             self.render_newcomment(content=content, error=error)
 
 
+class EditComment(Handler):
+    """Handles updating of comment entities and form validation."""
+    def get(self, comment_id, content="", error=""):
+        author = self.read_secure_cookie('user_id')
+        c = Comment.get_by_id(int(comment_id))
+        if not c:
+            self.error(404)
+            return
+
+        content = c.content
+        if author:
+            self.render('editcomment.html', content=content, error=error, author=int(author))
+        else:
+            self.render('editcomment.html', content=content, error=error)
+
+    def post(self, comment_id):
+        user_id = self.read_secure_cookie('user_id')
+        c = Comment.get_by_id(int(comment_id))
+        if not c:
+            self.error(404)
+            return
+
+        content = self.request.get('content')
+
+        if self.request.get('delete-comment'):
+            if c.user_id == int(user_id):
+                c.delete()
+                time.sleep(0.1)
+                self.redirect('/blog/%s'%c.blog_id)
+            else:
+                msg = "You do not own that comment."
+                self.redirect('/?msg=%s'%msg)
+        elif self.request.get('edit-comment'):
+            self.render('editcomment.html', content=c.content)
+        elif self.request.get('submit'):
+            if c.user_id == int(user_id):
+                if content:
+                    c.content = content
+                    c.put()
+                    time.sleep(0.1)
+                    self.redirect('/blog/%s' % c.blog_id)
+                else:
+                    error = "Content is required!"
+                    self.render('editcomment.html', content=content, error=error)
+            else:
+                error = "You do not own this comment!"
+                self.render('editcomment.html', content=content, error=error)
+        elif self.request.get('cancel'):
+            self.redirect('/blog/%s' % c.blog_id)
+
+
 class EditPost(Handler):
+    """Handles updating of blog entities and form validation."""
     def get(self, blog_id, subject="", content="", error=""):
         author = self.read_secure_cookie('user_id')
         b = Blog.get_by_id(int(blog_id))
+        if not b:
+            self.error(404)
+            return
+
         subject = b.subject
         content = b.content
         if author:
-            self.render("edit.html", blog_id=blog_id, subject=subject, content=content, error=error, author = author)
+            self.render('edit.html', blog_id=blog_id, subject=subject, content=content, error=error, author = int(author))
         else:
-            self.render("edit.html", blog_id=blog_id, subject=subject, content=content, error=error)
+            self.render('edit.html', blog_id=blog_id, subject=subject, content=content, error=error)
 
     def post(self, blog_id):
         b = Blog.get_by_id(int(blog_id))
-        subject = self.request.get("subject")
-        content = self.request.get("content")
+        if not b:
+            self.error(404)
+            return
+
+        subject = self.request.get('subject')
+        content = self.request.get('content')
         user_id = b.user_id
         if self.request.get('submit'):
             if user_id == self.read_secure_cookie('user_id'):
@@ -358,28 +436,29 @@ class EditPost(Handler):
 
                 else:
                     error = "Both subject and content are required!"
-                    self.render("edit.html", blog_id=blog_id, subject=subject, content=content, error=error)
+                    self.render('edit.html', blog_id=blog_id, subject=subject, content=content, error=error)
             else:
                 error = "You do not own this blog!"
-                self.render("edit.html", blog_id=blog_id, subject=subject, content=content, error=error)
+                self.render('edit.html', blog_id=blog_id, subject=subject, content=content, error=error)
         elif self.request.get('cancel'):
             self.redirect('/blog/%s' % b.key().id())
 
 
 class NewPost(Handler):
+    """Handles creation of blog entities and form validation."""
     def render_newpost(self, subject="", content="", error=""):
         author = self.read_secure_cookie('user_id')
         if author:
-            self.render("newpost.html", subject=subject, content=content, error=error, author=author)
+            self.render('newpost.html', subject=subject, content=content, error=error, author=author)
         else:
-            self.render("newpost.html", subject=subject, content=content, error=error)
+            self.redirect('/login')
 
     def get(self):
         self.render_newpost()
 
     def post(self):
-        subject = self.request.get("subject")
-        content = self.request.get("content")
+        subject = self.request.get('subject')
+        content = self.request.get('content')
         user_id = self.read_secure_cookie('user_id')
 
         if subject and content:
@@ -391,6 +470,7 @@ class NewPost(Handler):
             error = "Both subject and content are required!"
             self.render_newpost(subject=subject, content=content, error=error)
 
+
 app = webapp2.WSGIApplication([('/', HomePage),
                                ('/signup', Register),
                                ('/logout', Logout),
@@ -398,6 +478,7 @@ app = webapp2.WSGIApplication([('/', HomePage),
                                ('/welcome', Welcome),
                                ('/blog', MainBlogPage),
                                ('/comment/([0-9]+)', CommentHandler),
+                               ('/comment/edit/([0-9]+)', EditComment),
                                ('/blog/newpost', NewPost),
                                ('/blog/edit/([0-9]+)', EditPost),
                                ('/blog/([0-9]+)',BlogPage)], debug = True)
